Added new class 'Object'.
The object class should be the base for most of the Qube2D classes
because it provides a hierachy system that can store children of
different types and also allows to fetch children of a certain type
only. Furthermore, it can hold properties which are mapped to unique
string identifiers. The properties itself are Qube2D::Variant values
which can be numbers, strings and other Qube2D types.